1- create folder for zendesk utils and move the taskRouter logic to Helper class. 2- added the workerSid in warmTransfer action and validate in wrapUp event if the workerSid its the same to update the SyncDoc
