--- plugin-flex-ts-template-v2/src/feature-library/zendesk-extension/helpers/ZendeskExtensionService.ts
@@ -1,7 +1,0 @@
-import logger from '../../../utils/logger';
-
-class ZendeskExtensionService  {
-    
-}
-
-export default new ZendeskExtensionService();
