+++ plugin-flex-ts-template-v2/src/feature-library/zendesk-extension/utils/zendesk/ZendeskState.ts
